fix : 유틸 함수 정리
